Split yaml package into several files
